requirements.txt
openai
supabase
uuid
redis
python-dotenv
celery
tiktoken